Take the JEP 411 warning out back and shoot it
PL/Java has never, until now, implemented anything by reaching into
unexposed JDK innards. The need to do so now is fairly irksome.
Java, any language, is classic infrastructure. Other layers,
like this, are built atop it, and others in turn use those layers.
The idea that the language developers would arrogate to themselves
the act of sending an inappropriately low-level message directly
to ultimate users, insisting that the stack layers above cannot
intercept it and notify the higher-level users in terms that fit
the abstractions meaningful there, leaves an uneasy picture of
how a development team can begin to lose sight of who is providing
what to whom and why.