About the JEP 411 deprecation warnings
Because PL/Java 1.6 builds with --release 9, no @SuppressWarnings
annotations need to be added for JEP 411, as the deprecation markings
introduced in the JEP are not visible as of release 9.

I will turn on the maven-compiler-plugin's <showDeprecation> anyway,
as its default of false just seems wrong to me.

I'll create a branch from this point that has the @SuppressWarnings
annotations that would otherwise need to be added, so they will be
on record, but there is no need to merge them in for PL/Java 1.6.
May as well not clutter the code with them unnecessarily.