Supply advisory warnings appropriate to PL/Java
Having suppressed the boilerplate warning from Java itself (which
was going to spam standard error for every backend that starts PL/Java,
ending up in the server log if logging_collector is set), set about
supplying a more usefully worded warning at more strategically chosen
times, that will mention the PL/Java wiki/JEP-411 page for details on
the road map ahead.

The idea is to produce no more than one warning per backend, and only
at times when something PL/Java-administrative in nature is happening,
to maximize the chance that the person who occasions the warning is
a person it will mean something to.

So, deliver the warning:

1. when PL/Java itself is installed or upgraded (really, any time its
library is loaded by LOAD, directly or by the extension script, even if
it's the version already installed and nothing else happens).

2. when functions are declared or redeclared that use it (as detected
by the validator handler).

3. when pg_upgrade is run on a database that has PL/Java installed.

1 and 2 are only delivered if the Java major version is 12 or later.
The idea is that if a site tends to stick with long-term-support
Java releases, there will be plenty of time on 17 LTS for warnings,
and no need to deliver them while running 11 LTS. At another site
that follows along at a faster pace with non-LTS releases, start
with the warnings as soon as running anything later than 11.

3 is unconditional because under pg_upgrade a JVM doesn't get launched
to determine its version.

1 and 2 also defer the warning to the commit of a top-level
transaction that contained the triggering event(s). If the
transaction doesn't commit, the warning is squelched to avoid
adding noise to other messages possibly indicating something
went wrong.