Manage ContextClassLoader around PL/Java functions
This is done from the C side and, perhaps more controversially, using
JNI to poke a non-exposed field in java.lang.Thread. PL/Java had not
engaged in such tomfoolery before the JEP 411 warning situation required
it, and this is now the second instance. It bears mentioning that, from
ongoing discussion of JEP 411, the Java developers seem to be taking
the position that any project wishing to offer some kind of permission
enforcement post-JEP 411 is going to be more or less expected to do so
by poking at non-exposed innards using JNI, JVMTI, or the like. In some
post-1.6 PL/Java version, this kind of thing will have to become more
common.

In this case, it is simply to finally start setting the context loader
appropriately (I do think it's arguably a correctness issue, as
discussed in #361) while adding as little overhead as possible to the
hot path of invoking functions. This code will detect, and fall back to
not managing the context loader, if the target JRE lacks the expected
field. It can also be ineffective if an application has subclassed
Thread with a different behavior for getContextClassLoader. It seems
tolerable to leave such cases unhandled.

Because in PL/Java 1.6 the division of labor still has parameter and
return value conversions done from the C side before and after the
target method is invoked, and those conversions could involve Java UDTs
whose support methods are not bound to context loaders in their own
right, it makes sense to have the C code impose the target method's
context loader in advance of the parameter conversions, with restoration
only after return value conversion. Any UDT methods then consistently
see the target method's context loader, whether they are being invoked
'outside' it for parameter or return conversion, or 'inside' it during
PreparedStatement / ResultSet / SQLInput / SQLOutput operations.

Likewise, leave class initialization to happen, as it normally would,
on first use of a function from the class, when the right context
loader will be in place. It had been pulled up to validation time to
address a small quirk where OpenJ9's class resolution is so lazy it can
overlook missing-dependency problems HotSpot would flag, but it's
more fuss than the small quirk is worth.

This commit introduces some C code in the typedef-function-signature
pattern, rather than the typedef-function-pointer pattern prevalent
in PostgreSQL upstream, as discussed in [0]. Being valid C89/"ANSI",
and with PostgreSQL's documentation stipulating C89 or ANSI compilers
as early as PG 8.4 [1], this is expected to cause no problems in an
environment that can build any PG version PL/Java 1.6 supports.

[0] https://www.postgresql.org/message-id/615C847A.7070609%40anastigmatix.net
[1] https://www.postgresql.org/docs/8.4/source-format.html