Allow XSLT examples to use builtin impl or not
Add a 'builtin' argument (default true) requiring Java to find its
own in-the-box XSLT implementation; when false, an implementation
via the (now properly managed!) context classloader may be found;
for example, XSLT 3.0 transforms can be used if the schema class
path includes Saxon. When not using the builtin implementation,
treat as warnings, rather than errors, any failure to recognize
specific features/attributes the builin implementation is expected
to recognize.

It's appropriate for the static transformer factory s_tf to be
the builtin one.