Avoid arbitrary bits as a sentinel object ref
Because the Java API treats null as a meaningful value for
setContextClassLoader, something other than null must be used
as the sentinel "didn't set one" value. But choosing some arbitrary
int bits other than NULL to use as a sentinel pointer value
could be trouble on some architectures.

Assuming no logic errors, whatever sentinel value is used here
will never get passed to JNI and used as an object reference,
but should such an error exist, the results could be undefined
unless the value is an actual JNI global reference to a ClassLoader
instance. So, to play it safe, just make a dummy one of those
to use as the sentinel. If any sequence of mishaps leads Java to
use it as a class loader, it will throw null pointer exceptions,
but that beats undefined behavior.

Invocation_pushBootContext can just use null; its uses may precede
the allocation of the sentinel, and popBootContext doesn't try to
restore anything anyway.