Document and enforce MappedUDT/typlen-2 exclusion
Check at runtime whether a MappedUDT is being applied to a PostgreSQL
type declared with typlen of -2 (a variable-length, NUL-terminated
representation), and report an error if so. Explain in the developer
documentation why the implementation cannot support the use.

It already didn't work, but an explanatory error message is
a better way of not working.

Addresses #370.