Expose predefined category codes in BaseUDT
The PostgreSQL type 'category' is a bit awkward, because there are
several predefined, but other codes can be used for custom purposes.
So an enumeration of the predefined ones cannot be used as the type
of the 'category' annotation element, but it still can improve
readability when one of the predefined categories is what's wanted.

Also add a warning from the DDR processor if an upper-case category
code has been given (that's the range reserved for PostgreSQL's
predefined categories) but it doesn't correspond to any value of
the PredefinedCategory enum.