Fix mapping errors to java.time dates and times
The (documented, but not necessary) limitation in the mapping
of date infinity/-infinity to java.time.LocalDate was an
avoidable integer overflow. It only affected dates within 30 years
of infinity or -infinity.

Addresses #390.

In passing, solve the problem that the valid PostgreSQL time value
24:00:00.000000 is not valid to java.time.LocalTime or OffsetTime.
The Java classes will accept the preceding nanosecond, which is
still distinguishable from any other mapped PostgreSQL value
(as those have only microsecond resolution), so make that the
mapping of 24:00:00.