Create a 1.7-flavored branch
The complications imposed by Java's JEP 411 announcement will be
important to address for a 2.0 version, and yet there is significant
new development in progress that does not yet address that. For that,
a branch that follows the actively-released 1.6 branch, but does
not yet aspire to be 2.0, is appropriate.

For the time being, this branch will present highly explorative
work in progress. Building from this branch would only be recommended
for interested participants in PL/Java's development.