Easy warmup: less JNI in Invocation
Tweak invocation.c so the stack-allocated space provided by the caller
is used to save the prior state rather than to construct the new state.
This way, the current state can have a fixed address (currentInvocation
is a constant pointer) and can be covered by a single static
ByteBuffer that Invocation.java can read/write through without relying
on JNI methods.

As Invocation isn't a JDBC-specific concept or class, it has never
made much sense to have it in the .jdbc package. Move it to .internal.