Use SPI in the purely nested fashion it expects
In passing, fix a long-standing thinko in Invocation_popInvocation:
the memory context that was current on entry is stored in upperContext
of *this* Invocation, but popInvocation was 'restoring' the one that was
saved in the *previous* Invocation.

Also in passing, move the cleanEnqueuedInstances step later in the
pop sequence, improving its chance of seeing instances that could become
unreachable through the release of SPI contexts or the JNI local frame.