Also eliminate special SRF Invocation treatment
Without the special treatment, the instance of the Java class
Invocation, if any, that corresponds to the C Invocation, has its
lifetime simply bounded to that of the C Invocation, rather than
artificially extended across a sequence of SRF value-per-call
invocations. It is simpler, does not break any existing tests, and
is less likely to be violating PostgreSQL assumptions on correct
behavior.