Two caching tools: CacheMap and SwitchPointCache
CacheMap is a generic class useful for (possibly weak or soft)
canonicalizing caches of things that are identified by one or more
primitive values. (Writing the key values into a ByteBuffer avoids
the allocation involved in boxing them; however, the API as it
currently stands might be exceeding that cost with instantiation
of lambdas. It should eventually be profiled, and possibly revised
into a less tidy, but more efficient, form.)

SwitchPointCache is intended for lazily caching numerous values
of diverse types, groups of which can be associated with a single
SwitchPoint for purposes of invalidation.

As currently structured, the SwitchPoints (and their dependent
GuardWithTest nodes) do not get stored in static final fields;
this may limit HotSpot's ability to optimize them as fully as
it could if they did.