Adapter and org.postgresql.pljava.adt.spi package
Adapter is the abstract ancestor of all classes that implement
PostgreSQL datatypes for PL/Java, and the adt.spi package contains
classes that will be of use to datatype-implementing code:
in particular, Datum. PostgreSQL datums are only exposed
to Adapters, and the Adapter's job is to reliably convert between
the PostgreSQL type and some appropriate Java representation.

For some datatypes, there is a single or obvious appropriate Java
representation, and an Adapter may be provided that simply produces
that. For other datatypes, there may be no single obvious choice
of Java representation, either because there is no good match or
because there are several; an application might want to map types
to specialized classes available in some domain-specific library.
To serve those cases, Adapters can be defined in terms of
Adapter.Contract subinterfaces, which are simply functional interfaces
that document and expose the semantic components of the PostgreSQL
type. For example, a contract for PostgreSQL INTERVAL would expose
a 64-bit microseconds component, a 32-bit day count, and a 32-bit
month count. The division of responsibility is that the Adapter
encapsulates how to extract those components given a PostgreSQL
datum, but the contract fixes the semantics of what the components
are. It is then simple to use the Adapter, with any lambda that
conforms to the contract, to produce any desired Java representation
of the type.

Dummy versions of Attribute, RegClass, RegType, TupleDescriptor,
and TupleTableSlot break ground here on the model package, which
will consist of a set of classes modeling key PostgreSQL abstractions
and a useful subset of the PostgreSQL system catalogs.

RegType also implements java.sql.SQLType, making it usable in
(a suitable implementation of) JDBC to specify PostgreSQL types
precisely.

adt.spi.AbstractType needs the specialization() method that was
earlier added to internal.Function in anticipation of needing it
someday.