Package org.postgresql.pljava.adt: contracts
The org.postgresql.pljava.adt package contains 'contracts'
(subinterfaces of Adapter.Contract.Scalar or Adapter.Contract.Array),
which are functional interfaces that document and expose the exact
semantic components of PostgreSQL data types.

Adapters are responsible for the internal details of PostgreSQL's
representation that aren't semantically important, and code that
simply needs to construct some semantically faithful representation
of the type only needs to be concerned with the contract.