CharsetEncoding, CatalogObject, and a few more
CharsetEncoding is not really a catalog object (the available
encodings in PostgreSQL are hardcoded) but is exposed here as
a similar kind of object with useful operations, including
encoding and decoding using the corresponding Java codec when
known.

CatalogObject is, of course, the superinterface of all things
that really are catalog objects (identified by a classId, an objectId,
and rarely a subId). This commit brings in RegNamespace and RegRole
as needed for CatalogObject.Namespaced and CatalogObject.Owned.
RolePrincipal is a bridge between a RegRole and Java's Principal
interface.

CatalogObject.Factory is a service interface 'used' by the API
module, and will be 'provided' by the internals module to supply
the implementations of these things.