Expose MemoryContext and ResourceOwner in Java
These PostgreSQL notions will have to be available to Java code
for two reasons.

First, even code that has no business poking at them can still need
to know which one is current, to set an appropriate lifetime on
a Java object that corresponds to something in PostgreSQL allocated
in that context or registered to that owner. For that purpose, they
both will be exposed as subtypes of Lifespan, and the existing
PL/Java DualState class will be reworked to accept any Lifespan to
bound the validity of the native state.

Second, Adapter code could very well need to poke at such objects
(MemoryContexts, anyway): either to make a selected one current for
when allocating some object, or even to create and manage one.
Methods for that will not be exposed on MemoryContext or ResourceOwner
proper, but could be protected methods of Adapter, so that only
an Adapter can use them.