Now an Invocation is also a Lifespan
The current invocation can be the right Lifespan to specify for
a DualState that's guarding some object PostgreSQL passed in to
the call, which is expected to be good for as long as the call
is in progress.

In other, but related, news, Invocation can now return the
"upper executor" memory context: that is, whatever context was
current at entry, even if a later use of SPI changes the context
that is current.

It can appear tempting to eliminate the special treatment of PgSavepoint
in Invocation, and simply make it another DualState client, but because
of the strict nesting imposed on savepoints, keeping just the one
reference to the first one set suffices, and is more efficient.