Clients using Invocation.current as Lifespan
These DualState clients were previously passing the address of
the current invocation struct as their "resource owner", again from
the C code, passed along by the Java constructor. Again simplify
to call Invocation.current() right in the Java constructor and use
that as the Lifespan.

On a side note, the legacy Relation class included here (and its
legacy Tuple and TupleDesc) will naturally be among the first
candidates for retirement when this new model API is ready.