Eliminate last use of DualState.Key
Originally a hedge against coding mistakes during the introduction
of DualState for 1.5.1 (which had to support Java < 9), it is less
necessary now that the internals are behind JPMS encapsulation, and
the former checks for the cookie can be replaced with assertions that
the action is happening on the right thread. The CI tests run with
assertions enabled, so this should be adequate.