Adapt SQLXMLImpl to be creatable from Java
Fitting it into the new scheme is not entirely completed here;
for example, newReadable takes a Datum.Input parameter, but still
casts it internally to VarlenaWrapper.Input. Making it interoperate
with any Datum.Input may be a bit more work.

Likewise, newReadable with synthetic=true still encapsulates all
the knowledge of what datatypes there is synthetic-XML coverage
for and selecting the right VarlenaXMLRenderer for it (there's
that varlena-specificity again!). More of that should be moved
out of here and into an Adapter.

In passing, fix a couple typos in toString() methods, and add
a serviceable, if brute-force, getString() method to Synthetic.
It would be better for SyntheticXMLReader to gain the ability to
produce character-stream output efficiently, but until that
happens, there needs to be something for those moments when you
just want a string to look at and shouldn't have to fuss to get it.

For now, VarlenaWrapper.Input and .Stream still extend, and add small
features like toString(Object) to, DatumImpl. Later work can probably
migrate those bits so VarlenaWrapper will only contain logic specific
to varlenas.

An adt.spi interface Verifier is added, though Datum doesn't yet
expose any way to use it; in this commit, only one method accepting
Verifier.OfStream is added in DatumImpl.Input.Stream, the minimal
change needed to get things working.