Build out TupleTableSlot implementation
There are two flavors so far, Deformed and Heap. Deformed works
with whatever a real PostgreSQL TupleTableSlot can work with,
relying on the PostgreSQL implementation to 'deform' it into
separate datum and isnull arrays. (That doesn't have to be a
PostgreSQL 'virtual' TupleTableSlot; it can do the deforming
independently of the type of slot. When the time comes to
implement the reverse direction and produce tuples, a virtual
slot will be the way to go for that, using the PostgreSQL C code
to 'form' it once populated.)

The Heap flavor knows enough about that PostgreSQL tuple format
to 'deform' it in Java without the JNI calls (except where some
out-of-line value has to be mapped, or for varlena values until
VarlenaWrapper sheds more of its remaining JNI-centricity). The
Heap implementation does not yet do anything clever to memoize
the offsets into the tuple, which makes the retrieval of all
the tuple's values an O(n^2) proposition; there is a
low-hanging-fruit optimization opportunity there. For now, it gets
the job done.

It might be interesting to see how the two flavors compare on
typical heap tuples: Deformed, making more JNI calls but relying
on PostgreSQL's fast native deforming, or Heap, which can avoid
more JNI calls, and also avoids deforming something into a fresh
native memory allocation if the only thing it will be used for is
to immediately construct some Java object.

The Heap flavor can do one thing the Deformed flavor definitely
cannot: it can operate on heap-tuple-formatted contents of an
arbitrary Java byte buffer, which in theory might not even be
backed by native memory. (Again, for now, this is slightly science
fiction where varlena values are concerned, because VarlenaWrapper
retains a lot of its native dependencies. A ByteBuffer "heap tuple"
with varlenas in it will have to be native-backed for now.) The
selection of the DualState guard by heapTupleGetLightSlot() is
currently more hardcoded than that would suggest; it assumes the
buffer is mapping memory that can be heap_free_tuple'd.

The 'light' in heapTupleGetLightSlot really means that there isn't
an underlying PostgreSQL TupleTableSlot constructed.

The whole business of how to apply and use DualState guards on these
things still needs more attention.

There is also Heap.Indexed, which is the thing needed for arrays.
When the element type is fixed-length, it achieves O(1) access
(plus null-bitmap processing if there are nulls). It uses a "count
preceding null bits ahead of time" strategy that could also easily
be adopted in Heap.

A NullableDatum flavor is also needed, which would be the thing for
mapping (as one prominent example) function-call arguments.

The HeapTuples8 and HeapTuples4 classes at the end are scaffolding
and ought to be factored out into something with a decent API, as
hinted at in the comment preceding them.

A Heap instance still inherits the values/nulls array fields used
in the deformed case, without (at present) making any use of them.
It is possible some use could be made (as, again, an underlying PG
TupleTableSlot could be used in deforming a heap tuple), but it's
also possible that won't ever be needed, and the class could be
refactored to a simpler form.