Build out the CatalogObject implementation
Here's how this is going to work.

The "exists because mentioned" aspect of a CatalogObject is
a lightweight operation, just caching/returning a singleton with
the mentioned values of classId/objId/(subId?).

For a bare CatalogObject (objId unaccompanied by classId), that's
all there is. But for any CatalogObject.Addressed subtype, the
classId and objId together identify a tuple in a particular system
catalog (or, that is, identify a tuple that could exist in that
catalog). And the methods on the Java class that return information
about the object get the information by fetching attributes from
that tuple, then constructing whatever the Java representation
will be.

Not to duplicate the work of fetching (the tuple itself, and then
an attribute from the tuple) and constructing the Java result, an
instance will have an array of SwitchPointCache-managed "slots"
that will cache, lazily, the constructed results. Five of those
slots have their indices standardized right here in CatalogObjectImpl,
to account for the name, namespace, owner, and ACL of objects that
have those things. Slot 0 is for the tuple itself.

When an uncached value is requested, the "computation method" set up
for that slot will execute (always on the PG thread, so it can
interact with PostgreSQL with no extra ceremony). Most computation
methods will begin by calling cacheTuple() to obtain the tuple
itself from slot 0, and then will fetch the wanted attribute from it
and construct the result. The computation method for cacheTuple(),
in turn, will obtain the tuple if that hasn't happened yet, usually
from the PostgreSQL syscache. We copy it to a long-lived memory
context where we can keep it until its invalidation.

The most common way the cacheTuple is fetched is by a one-argument
syscache search by the object's Oid. When that is all that is needed,
the Java class need only implement cacheId() to return the number
of the PostgreSQL syscache to search in. For exceptional cases
(attributes, for example, require a two-argument syscache search),
a class should just provide its own cacheTuple computation method.

The slots for an object are associated with a Java SwitchPoint,
and the mapping from the object to its associated SwitchPoint
is a function supplied to the SwitchPointCache.Builder. Some
classes, such as RegClass and RegType, will allocate a SwitchPoint
per object, and can be selectively invalidated. Otherwise, by
default, the s_globalPoint declared here can be used, which will
invalidate all values of all slots depending on it.