RegClass and RegType have a special relationship
They are the two CatalogObjects with tupleDescriptor() methods.

You can get strictly more tuple descriptors by asking RegType;
a RegType.Blessed can give you a tuple descriptor that has been
interned in the PostgreSQL typcache and corresponds to nothing
in the system catalogs. But whenever a RegType t is an ordinary
cataloged composite type or the row type of a cataloged relation,
then there is a RegClass c such that c == t.relation() and
t == c.type(), and you will get the same tuple descriptor from
the tupleDescriptor() method of either c or t.

In all but one such case, c delegates to c.type().tupleDescriptor()
and lets the RegType do the work, obtaining the descriptor from
the PG typcache.

The one exception is when the tuple descriptor for pg_class itself
is wanted, in which case the RegClass does the work, obtaining the
descriptor from the PG relcache, and RegType delegates to it for
that one exceptional case. The reason is that RegClass will see
the first request for the pg_class tuple descriptor, and before that
is available, c.type() can't be evaluated.

In either case, whichever class looked it up, a cataloged tuple
descriptor is always stored on the RegClass instance, and RegClass
will be responsible for its invalidation if the relation is altered.
(A RegType.Blessed has its own field for its tuple descriptor,
because there is no corresponding RegClass for one of those.)

Because of this close connection between RegClass and RegType,
the methods RegClass.type() and RegType.relation() use a handshake
protocol to ensure that, whenever either method is called, not only
does it cache the result, but its counterpart for that result instance
caches the reverse result, so the connection can later be traversed
in either direction with no need for a lookup by oid.

In the static initializer pattern introduced here, the handful of
SwitchPointCache slots that are predefined in CatalogObject.Addressed
are added to, by starting an int index at Addressed.NSLOTS,
incrementing it to initialize additional slot index constants, then
using its final value to define a new NSLOTS that shadows the original.