Attribute, naturally, has complications
An Attribute is most often obtained from a TupleDescriptor
(in this API, that's how it's done), and the TupleDescriptor
can supply a version of Attribute's tuple directly; no need
to look it up anywhere else. That copy, however, cuts off
at ATTRIBUTE_FIXED_PART_SIZE bytes. The most commonly needed
attributes of Attribute are found there, but for others beyond
that cutoff, the full tuple has to be fetched from the syscache.

So AttributeImpl has the normal SLOT_TUPLE slot, used for the
rarely-needed full tuple, and also its own SLOT_PARTIALTUPLE,
for the truncated version obtained from the containing tuple
descriptor. Most computation methods will fetch from the partial
one, with the full one referred to only by the ones that need it.

It doesn't end there. A few critical Attribute properties, byValue,
alignment, length, and type/typmod, are needed to successfully fetch
values from a TupleTableSlotImpl.Heap. So Attribute cannot use that
API to fetch those values. For those, it must hardcode their actual
offsets and sizes in the raw ByteBuffer that the containing tuple
descriptor supplies, and fetch them directly. So there is also
a SLOT_RAWBUFFER.

This may sound more costly in space than it is. The raw buffer,
of course, is just a ByteBuffer sliced off and sharing the larger
one in the TupleDescriptor, and the partial tuple is just a
TupleTableSlot instance built over that. The full tuple is another
complete copy, but only fetched when those less-commonly-needed
attributes are requested.

With those key values obtained from the raw buffer, the Attribute's
name does not require any such contortions, and can be fetched using
the civilized TupleTableSlot API, except it can't be done by name,
so the attribute number is used for that one.

An AttributeImpl.Transient holds a direct reference to
the TupleDescriptor it came from, which its containingTupleDescriptor()
method returns. An AttributeImpl.Cataloged does not, and instead holds
a reference to the RegClass for which it is defined in the system
catalogs, and containingTupleDescriptor() delegates to tupleDescriptor()
on that. If the relation has been altered, that could return an updated
new tuple descriptor.