Here's a simple complete one, a Database object
Which means also bringing in a GrantAdapter (needed for the full
implementation of the CatalogObject.AccessControlled subtypes) and
EncodingAdapter (needed for this and RegCollation, anyway).

This also has a static final CURRENT field for convenient access
to the instance representing the current database, bringing one more
native method to obtain the value.