Pacify AppVeyor
The "MinGW PG 12" cases were succeeding because the MinGW PacMan
environment is really supplying PG 14. Update the "PG" variable
for those cases; it's ignored, but may as well reflect (this
moment's) reality.

The MSVC cases really used PG 12 and failed because of a struct
field not present before 13. For now, just consider 13 to be minimum
for this work in progress. There are several things that would have to
be made conditional to work on PG < 13. So far, only one #include change
seems necessary to work on 13.