Give up static check of OFFSET_ArrayType_DIMS
It appears that ARR_DIMS is defined in such a way as to preclude
its use in a constant expression, at least according to clang.

In this commit, it is moved to a runtime check. The check of
SIZEOF_ArrayType_DIM is not given up on yet; clang did not report
that as an error in the last run.