More work on array adapters
A Contract.Array now has a type parameter for the expected element
adapter type, which can range over the reference-returning As<T,?>
and any of the primitive-returning AsFoo<?> types, so it is now
straightforward to implement an array contract and adapter over
an array whose element type will be mapped to a Java primitive.

The ArgModeAdapter goes away, as suggested in its javadoc, now that
an adapter for list-of-ArgMode can be easily thrown together from
an INT1 element adapter and an anonymous array contract.

In passing, also reorder ArrayAdapter constructor arguments, putting
the element adapter first. In constructing an ArrayAdapter with an
in-place anonymous array contract, the element adapter was becoming
hard to see when it came at the end of all that.

A reference-returning (As<T,?>) adapter now has a method to allocate
a T[] array, relieving the usual Java headache of getting the expected
compile-time typing on an array. AsFlatList.nullsIncludedCopy now uses
that and returns an unmodifiable list.

A bunch of javadoc filled in.