Fail validation if TRANSFORM FOR TYPE declared
The TRANSFORM FOR TYPE mechanism is only meaningful for a PL
whose handler function itself does the work of looking up
and applying the transforms.

https://www.postgresql.org/message-id/61FDBCFE.3090800%40anastigmatix.net

This PL doesn't, so say so when validating a function declaration
that includes a TRANSFORM clause.

It would arguably be better to say so even earlier, at an attempted
CREATE TRANSFORM for the PL, but PostgreSQL does not currently apply
any PL-specific validator function at that time.